Decoder Wheel MVP
A functional system to create pseudo-random ordering of the code
wheel key, and a webpage to allow the printing of the code wheel,
with instructions for its use.  If anybody knows the name of this
algorithm, which is some sort of variant of the Vigenere Cipher, I
please email me so I can read up on the particulars of the cipher
and see if my implementation is working as well as it could or not.